cleanup(core): extract legacy dependsOn warning into its own file
